generate_teams_user_aad_token replaced to test utils
